moved all code from probe_installer to setup.py
